+++ CHANGELOG.md
@@ -49,6 +49,13 @@ All notable changes to this project will be documented in this file.
 
 - Created a function to handle to various Capacity Tracker non-residential cleaning steps.
 
+- Generalised the functions to create and update a filtering column
+
+- Created a function to handle to various Capacity Tracker non-residential cleaning steps.
+
+- Added NHS Capacity Tracker filter to remove repeated submissions based on provider total post being the same
+  for too long.
+
 ### Changed
 - Migrated Polars validation scripts over to use PointBlank (compatible with >= Python 3.11), so far:
   - locations_raw